Support for HLS source type
This pull request simply changes the regex that checks the source.type, allowing for the HLS (.m3u8) mime type "application/x-mpegURL" to be used.

Currently if you are trying to stream in HLS .m3u8 you can specify the type as dash "application/dash+xml" and it works just fine... even though it's the wrong mime type. This pull requests allows to specify the correct mime type for HLS streams.

Tested with a private streams and it works.